refactor(server): split runtime bootstrap helpers
